+++ src/Fantomas.Core.Tests/ClassTests.fs
@@ -1073,3 +1073,77 @@ type Foo() =
         and set (value) = // comment set
             myInternalValue <- value
 """
+
+[<Test>]
+let ``preserve override keyword in read property, 2221`` () =
+    formatSourceString
+        false
+        """
+type StreamHelper() =
+    inherit Stream()
+
+    override x.ReadAsync (dst, offset, count, tok) = 
+        ()
+    override x.WriteAsync (dst, offset, count, tok) = 
+        ()
+    override x.Flush () = ()
+    override x.Seek(offset:int64, origin:SeekOrigin) =
+        ()
+    override x.SetLength(value:int64) =
+        ()
+    override x.Read(dst, offset, count) = 
+        ()           
+    override x.Write(src, offset, count) = 
+        ()
+    override x.ReadByte() =
+        ()
+    override x.WriteByte item =
+        ()
+    override x.CanRead 
+        with get() =
+            true
+    override x.CanSeek
+        with get() =
+            false
+    override x.CanWrite
+        with get() =
+            true
+    override x.Length
+        with get() =
+            1
+    override x.Position
+        with get() =
+            1
+        and set value =
+            1
+    override x.Dispose disposing =
+        ()
+"""
+        config
+    |> prepend newline
+    |> should
+        equal
+        """
+type StreamHelper() =
+    inherit Stream()
+
+    override x.ReadAsync(dst, offset, count, tok) = ()
+    override x.WriteAsync(dst, offset, count, tok) = ()
+    override x.Flush() = ()
+    override x.Seek(offset: int64, origin: SeekOrigin) = ()
+    override x.SetLength(value: int64) = ()
+    override x.Read(dst, offset, count) = ()
+    override x.Write(src, offset, count) = ()
+    override x.ReadByte() = ()
+    override x.WriteByte item = ()
+    override x.CanRead = true
+    override x.CanSeek = false
+    override x.CanWrite = true
+    override x.Length = 1
+
+    override x.Position
+        with get () = 1
+        and set value = 1
+
+    override x.Dispose disposing = ()
+"""