+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 * Idempotency problem with ParsedHashDirective in signature file. [#2258](https://github.com/fsprojects/fantomas/issues/2258)
+* Keyword 'override' gets changed to 'member'. [#2221](https://github.com/fsprojects/fantomas/issues/2221)
 
 ## [5.0.0-alpha-007] - 2022-05-16